[feat] 답변 생성 권한 확인 구현

+++ server/overflow/src/main/java/com/preproject/overflow/answer/service/AnswerService.java
@@ -25,13 +25,12 @@ public class AnswerService {
         this.questionService = questionService;
     }
     // 답변 생성 service
-    public Answer createAnswer(Answer answer) {
+    public Answer createAnswer(Answer answer, Member member) {
 
         // Question question = answer.getQuestion();
         // Member member = answer.getMember();
         // 위에를 아래처럼 수정, 위는 아이디만 넣는 것 같고 아래는 아이디값을 가지고 존재하는지 확인 후 객체를 넣는다.
         Question question = questionService.findVerifyQuestion(answer.getQuestionId());
-        Member member = memberService.findVerifiedMember(answer.getMemberId());
 
         answer.setQuestion(question);
         answer.setMember(member);
@@ -40,12 +39,12 @@ public class AnswerService {
         return answerRepository.save(answer);
     }
     // 답변 수정 service
-    public Answer updateAnswer(Answer answer, long memberId) {
+    public Answer updateAnswer(Answer answer, Member member) {
         // 존재하는 답변인지 찾은 후 업데이트
         Answer findAnswer = findVerifiedAnswer(answer.getAnswerId());
         // 답변을 작성한 사람과 수정하려는 사람이 일치하는지 확인
-        if(!(findAnswer.getMember().getMemberId() == memberId)) {
-            throw new BusinessLogicException(ExceptionCode.MEMBER_NOT_FOUND);
+        if(member.getMemberId() != findAnswer.getMember().getMemberId()){
+            throw new BusinessLogicException(ExceptionCode.MEMBER_PERMISSION_DENIED);
         }
 
         Optional.ofNullable(answer.getText())
@@ -63,13 +62,12 @@ public class AnswerService {
         return answerRepository.findAll();
     }
     // 답변 삭제 service
-    public void deleteAnswer(long answerId) {
-        Answer answer = new Answer();
-        answer.setAnswerId(answerId);
+    public void deleteAnswer(long answerId, Member member) {
+        Answer answer = findVerifiedAnswer(answerId);
 
-//        if(!(answer.getMember().getMemberId() == memberId)) {
-//            throw new BusinessLogicException(ExceptionCode.MEMBER_NOT_FOUND);
-//        }
+        if(member.getMemberId() != answer.getMember().getMemberId()){
+            throw new BusinessLogicException(ExceptionCode.MEMBER_PERMISSION_DENIED);
+        }
 
         answerRepository.delete(answer);
     }
